refactor to new structure
